mybatis整合 Signed-off-by: caihe <caihe@gintong.com>

+++ phoenix-knowledge-provider/src/main/java/com/ginkgocap/ywxt/knowledge/dao/knowledgecategory/KnowledgeCategoryDAO.java
@@ -16,11 +16,13 @@ public interface KnowledgeCategoryDAO {
 			long column_id, String pic_path);
 
 	int deleteKnowledgeRCategory(long[] knowledgeids, long categoryid);
+
 	/**
 	 * 查询栏目下知识个数
+	 * 
 	 * @param id
 	 * @return
 	 */
-	long countByKnowledgeCategoryId(long id);
+	int countByKnowledgeCategoryId(long id);
 
 }
